+++ tests/templates.py
@@ -2534,3 +2534,687 @@ Locale: US
 Description length: 72
 
 """
+
+
+@templates.register
+class _TestReadmeHelp(BaseTemplate):
+    @property
+    def template(self) -> str:
+        return """\
+|
+
+.. image:: https://raw.githubusercontent.com/jshwi/docsig/master/docs/static/docsig.svg
+   :alt: docsig logo
+   :width: 50%
+   :align: center
+
+|
+
+.. image:: https://img.shields.io/badge/License-MIT-yellow.svg
+   :target: https://opensource.org/licenses/MIT
+   :alt: License
+.. image:: https://img.shields.io/pypi/v/docsig
+   :target: https://pypi.org/project/docsig/
+   :alt: PyPI
+.. image:: https://github.com/jshwi/docsig/actions/workflows/build.yaml/badge.svg
+   :target: https://github.com/jshwi/docsig/actions/workflows/build.yaml
+   :alt: CI
+.. image:: https://github.com/jshwi/docsig/actions/workflows/codeql-analysis.yml/badge.svg
+   :target: https://github.com/jshwi/docsig/actions/workflows/codeql-analysis.yml
+   :alt: CodeQL
+.. image:: https://results.pre-commit.ci/badge/github/jshwi/docsig/master.svg
+   :target: https://results.pre-commit.ci/latest/github/jshwi/docsig/master
+   :alt: pre-commit.ci status
+.. image:: https://codecov.io/gh/jshwi/docsig/branch/master/graph/badge.svg
+   :target: https://codecov.io/gh/jshwi/docsig
+   :alt: codecov.io
+.. image:: https://readthedocs.org/projects/docsig/badge/?version=latest
+   :target: https://docsig.readthedocs.io/en/latest/?badge=latest
+   :alt: readthedocs.org
+.. image:: https://img.shields.io/badge/python-3.8-blue.svg
+   :target: https://www.python.org/downloads/release/python-380
+   :alt: python3.8
+.. image:: https://img.shields.io/badge/code%20style-black-000000.svg
+   :target: https://github.com/psf/black
+   :alt: Black
+.. image:: https://img.shields.io/badge/%20imports-isort-%231674b1?style=flat&labelColor=ef8336
+   :target: https://pycqa.github.io/isort/
+   :alt: isort
+.. image:: https://img.shields.io/badge/%20formatter-docformatter-fedcba.svg
+   :target: https://github.com/PyCQA/docformatter
+   :alt: docformatter
+.. image:: https://img.shields.io/badge/linting-pylint-yellowgreen
+   :target: https://github.com/PyCQA/pylint
+   :alt: pylint
+.. image:: https://img.shields.io/badge/security-bandit-yellow.svg
+   :target: https://github.com/PyCQA/bandit
+   :alt: Security Status
+.. image:: https://snyk.io/test/github/jshwi/docsig/badge.svg
+   :target: https://snyk.io/test/github/jshwi/docsig/badge.svg
+   :alt: Known Vulnerabilities
+.. image:: https://snyk.io/advisor/python/docsig/badge.svg
+   :target: https://snyk.io/advisor/python/docsig
+   :alt: docsig
+
+Check signature params for proper documentation
+-----------------------------------------------
+
+Supports reStructuredText (``Sphinx``), ``NumPy``, and ``Google``
+
+Contributing
+------------
+If you are interested in contributing to ``docsig`` please read about contributing `here <https://github.com/jshwi/docsig/blob/master/CONTRIBUTING.md>`__
+
+Installation
+------------
+
+.. code-block:: console
+
+    $ pip install docsig
+
+Usage
+-----
+
+Commandline
+***********
+
+.. code-block:: console
+
+    usage: docsig [-h] [-v] [-c] [-D] [-o] [-p] [-P] [-i] [-a] [-k] [-n] [-s STR]
+                    [path [path ...]]
+
+    Check signature params for proper documentation
+
+    positional arguments:
+      path                          directories or files to check
+
+    optional arguments:
+      -h, --help                    show this help message and exit
+      -v, --version                 show program's version number and exit
+      -c, --check-class             check class docstrings
+      -D, --check-dunders           check dunder methods
+      -o, --check-overridden        check overridden methods
+      -p, --check-protected         check protected functions and classes
+      -P, --check-property-returns  check property return values
+      -i, --ignore-no-params        ignore docstrings where parameters are not documented
+      -a, --ignore-args             ignore args prefixed with an asterisk
+      -k, --ignore-kwargs           ignore kwargs prefixed with two asterisks
+      -n, --no-ansi                 disable ansi output
+      -s STR, --string STR          string to parse instead of files
+
+Options can also be configured with the pyproject.toml file
+
+If you find the output is too verbose then the report can be configured to display a summary
+
+.. code-block:: toml
+
+    [tool.docsig]
+    check-dunders = false
+    check-overridden = false
+    check-protected = false
+    summary = true
+    disable = [
+        "E101",
+        "E102",
+        "E103",
+    ]
+    target = [
+        "E102",
+        "E103",
+        "E104",
+    ]
+
+API
+***
+
+.. code-block:: python
+
+    >>> from docsig import docsig
+
+.. code-block:: python
+
+    >>> string = \"\"\"
+    ... def function(param1, param2, param3) -> None:
+    ...     '''Summary for passing docstring...
+    ...
+    ...     Explanation for passing docstring...
+    ...
+    ...     :param param1: About param1.
+    ...     :param param2: About param2.
+    ...     :param param3: About param3.
+    ...     '''
+    ...     \"\"\"
+    >>> docsig(string=string)
+    0
+
+.. code-block:: python
+
+    >>> string = \"\"\"
+    ... def function(param1, param2) -> None:
+    ...     '''Summary for failing docstring...
+    ...
+    ...     Explanation for failing docstring...
+    ...
+    ...     :param param1: About param1.
+    ...     :param param2: About param2.
+    ...     :param param3: About param3.
+    ...     '''
+    ... \"\"\"
+    >>> docsig(string=string)
+    2
+    -
+    def function(✓param1, ✓param2, ✖None) -> ✓None:
+        \"\"\"
+        :param param1: ✓
+        :param param2: ✓
+        :param param3: ✖
+        \"\"\"
+    <BLANKLINE>
+    E102: includes parameters that do not exist
+    <BLANKLINE>
+    1
+
+A full list of checks can be found `here <https://docsig.readthedocs.io/en/latest/docsig.html#docsig-messages>`__
+
+Classes
+#######
+Checking a class docstring is not enabled by default, as this check is opinionated, and won't suite everyone
+
+This check will check documentation of ``__init__`` under the class docstring, and not under ``__init__`` itself
+
+.. code-block:: python
+
+    >>> string = \"\"\"
+    ... class Klass:
+    ...     '''Summary for failing docstring...
+    ...
+    ...     Explanation for failing docstring...
+    ...
+    ...     :param param1: About param1.
+    ...     :param param2: About param2.
+    ...     :param param3: About param3.
+    ...     '''
+    ...     def __init__(self, param1, param2) -> None:
+    ...         pass
+    ... \"\"\"
+    >>> docsig(string=string, check_class=True)
+    11 in Klass
+    -----------
+    class Klass:
+        \"\"\"
+        :param param1: ✓
+        :param param2: ✓
+        :param param3: ✖
+        \"\"\"
+    <BLANKLINE>
+        def __init__(✓param1, ✓param2, ✖None) -> ✓None:
+    <BLANKLINE>
+    E102: includes parameters that do not exist
+    <BLANKLINE>
+    1
+
+Checking class docstrings can be permanently enabled in the pyproject.toml file
+
+.. code-block:: toml
+
+    [tool.docsig]
+    check-class = true
+
+pre-commit
+**********
+
+``docsig`` can be used as a `pre-commit <https://pre-commit.com>`_ hook
+
+It can be added to your .pre-commit-config.yaml as follows:
+
+.. code-block:: yaml
+
+    repos:
+      - repo: https://github.com/jshwi/docsig
+        rev: v0.30.0
+        hooks:
+          - id: docsig
+            args:
+              - "--check-class"
+              - "--check-dunders"
+              - "--check-overridden"
+              - "--check-protected"
+              - "--summary"
+"""
+
+    @property
+    def expected(self) -> str:
+        return """\
+|
+
+.. image:: https://raw.githubusercontent.com/jshwi/docsig/master/docs/static/docsig.svg
+   :alt: docsig logo
+   :width: 50%
+   :align: center
+
+|
+
+.. image:: https://img.shields.io/badge/License-MIT-yellow.svg
+   :target: https://opensource.org/licenses/MIT
+   :alt: License
+.. image:: https://img.shields.io/pypi/v/docsig
+   :target: https://pypi.org/project/docsig/
+   :alt: PyPI
+.. image:: https://github.com/jshwi/docsig/actions/workflows/build.yaml/badge.svg
+   :target: https://github.com/jshwi/docsig/actions/workflows/build.yaml
+   :alt: CI
+.. image:: https://github.com/jshwi/docsig/actions/workflows/codeql-analysis.yml/badge.svg
+   :target: https://github.com/jshwi/docsig/actions/workflows/codeql-analysis.yml
+   :alt: CodeQL
+.. image:: https://results.pre-commit.ci/badge/github/jshwi/docsig/master.svg
+   :target: https://results.pre-commit.ci/latest/github/jshwi/docsig/master
+   :alt: pre-commit.ci status
+.. image:: https://codecov.io/gh/jshwi/docsig/branch/master/graph/badge.svg
+   :target: https://codecov.io/gh/jshwi/docsig
+   :alt: codecov.io
+.. image:: https://readthedocs.org/projects/docsig/badge/?version=latest
+   :target: https://docsig.readthedocs.io/en/latest/?badge=latest
+   :alt: readthedocs.org
+.. image:: https://img.shields.io/badge/python-3.8-blue.svg
+   :target: https://www.python.org/downloads/release/python-380
+   :alt: python3.8
+.. image:: https://img.shields.io/badge/code%20style-black-000000.svg
+   :target: https://github.com/psf/black
+   :alt: Black
+.. image:: https://img.shields.io/badge/%20imports-isort-%231674b1?style=flat&labelColor=ef8336
+   :target: https://pycqa.github.io/isort/
+   :alt: isort
+.. image:: https://img.shields.io/badge/%20formatter-docformatter-fedcba.svg
+   :target: https://github.com/PyCQA/docformatter
+   :alt: docformatter
+.. image:: https://img.shields.io/badge/linting-pylint-yellowgreen
+   :target: https://github.com/PyCQA/pylint
+   :alt: pylint
+.. image:: https://img.shields.io/badge/security-bandit-yellow.svg
+   :target: https://github.com/PyCQA/bandit
+   :alt: Security Status
+.. image:: https://snyk.io/test/github/jshwi/docsig/badge.svg
+   :target: https://snyk.io/test/github/jshwi/docsig/badge.svg
+   :alt: Known Vulnerabilities
+.. image:: https://snyk.io/advisor/python/docsig/badge.svg
+   :target: https://snyk.io/advisor/python/docsig
+   :alt: docsig
+
+Check signature params for proper documentation
+-----------------------------------------------
+
+Supports reStructuredText (``Sphinx``), ``NumPy``, and ``Google``
+
+Contributing
+------------
+If you are interested in contributing to ``docsig`` please read about contributing `here <https://github.com/jshwi/docsig/blob/master/CONTRIBUTING.md>`__
+
+Installation
+------------
+
+.. code-block:: console
+
+    $ pip install docsig
+
+Usage
+-----
+
+Commandline
+***********
+
+.. code-block:: console
+
+    usage: docsig [-h] [-v] [-c] [-D] [-o] [-p] [-P] [-i] [-a] [-k] [-n] [-S] [-s STR] path [path ...]
+
+    Check signature params for proper documentation
+
+    positional arguments:
+      path                          directories or files to check
+
+    optional arguments:
+      -h, --help                    show this help message and exit
+      -v, --version                 show program's version number and exit
+      -c, --check-class             check class docstrings
+      -D, --check-dunders           check dunder methods
+      -o, --check-overridden        check overridden methods
+      -p, --check-protected         check protected functions and classes
+      -P, --check-property-returns  check property return values
+      -i, --ignore-no-params        ignore docstrings where parameters are not documented
+      -a, --ignore-args             ignore args prefixed with an asterisk
+      -k, --ignore-kwargs           ignore kwargs prefixed with two asterisks
+      -n, --no-ansi                 disable ansi output
+      -S, --summary                 print a summarised report
+      -s STR, --string STR          string to parse instead of files
+
+Options can also be configured with the pyproject.toml file
+
+If you find the output is too verbose then the report can be configured to display a summary
+
+.. code-block:: toml
+
+    [tool.docsig]
+    check-dunders = false
+    check-overridden = false
+    check-protected = false
+    summary = true
+    disable = [
+        "E101",
+        "E102",
+        "E103",
+    ]
+    target = [
+        "E102",
+        "E103",
+        "E104",
+    ]
+
+API
+***
+
+.. code-block:: python
+
+    >>> from docsig import docsig
+
+.. code-block:: python
+
+    >>> string = \"\"\"
+    ... def function(param1, param2, param3) -> None:
+    ...     '''Summary for passing docstring...
+    ...
+    ...     Explanation for passing docstring...
+    ...
+    ...     :param param1: About param1.
+    ...     :param param2: About param2.
+    ...     :param param3: About param3.
+    ...     '''
+    ...     \"\"\"
+    >>> docsig(string=string)
+    0
+
+.. code-block:: python
+
+    >>> string = \"\"\"
+    ... def function(param1, param2) -> None:
+    ...     '''Summary for failing docstring...
+    ...
+    ...     Explanation for failing docstring...
+    ...
+    ...     :param param1: About param1.
+    ...     :param param2: About param2.
+    ...     :param param3: About param3.
+    ...     '''
+    ... \"\"\"
+    >>> docsig(string=string)
+    2
+    -
+    def function(✓param1, ✓param2, ✖None) -> ✓None:
+        \"\"\"
+        :param param1: ✓
+        :param param2: ✓
+        :param param3: ✖
+        \"\"\"
+    <BLANKLINE>
+    E102: includes parameters that do not exist
+    <BLANKLINE>
+    1
+
+A full list of checks can be found `here <https://docsig.readthedocs.io/en/latest/docsig.html#docsig-messages>`__
+
+Classes
+#######
+Checking a class docstring is not enabled by default, as this check is opinionated, and won't suite everyone
+
+This check will check documentation of ``__init__`` under the class docstring, and not under ``__init__`` itself
+
+.. code-block:: python
+
+    >>> string = \"\"\"
+    ... class Klass:
+    ...     '''Summary for failing docstring...
+    ...
+    ...     Explanation for failing docstring...
+    ...
+    ...     :param param1: About param1.
+    ...     :param param2: About param2.
+    ...     :param param3: About param3.
+    ...     '''
+    ...     def __init__(self, param1, param2) -> None:
+    ...         pass
+    ... \"\"\"
+    >>> docsig(string=string, check_class=True)
+    11 in Klass
+    -----------
+    class Klass:
+        \"\"\"
+        :param param1: ✓
+        :param param2: ✓
+        :param param3: ✖
+        \"\"\"
+    <BLANKLINE>
+        def __init__(✓param1, ✓param2, ✖None) -> ✓None:
+    <BLANKLINE>
+    E102: includes parameters that do not exist
+    <BLANKLINE>
+    1
+
+Checking class docstrings can be permanently enabled in the pyproject.toml file
+
+.. code-block:: toml
+
+    [tool.docsig]
+    check-class = true
+
+pre-commit
+**********
+
+``docsig`` can be used as a `pre-commit <https://pre-commit.com>`_ hook
+
+It can be added to your .pre-commit-config.yaml as follows:
+
+.. code-block:: yaml
+
+    repos:
+      - repo: https://github.com/jshwi/docsig
+        rev: v0.30.0
+        hooks:
+          - id: docsig
+            args:
+              - "--check-class"
+              - "--check-dunders"
+              - "--check-overridden"
+              - "--check-protected"
+              - "--summary"
+"""
+
+
+EXECUTABLE = '''
+"""
+docsig._main
+============
+
+Contains package entry point.
+"""
+from argparse import HelpFormatter as _HelpFormatter
+from pathlib import Path as _Path
+
+from argparse import ArgumentParser as _ArgumentParser
+
+
+NAME = __name__.split(".", maxsplit=1)[0]
+
+
+class _Parser(_ArgumentParser):
+    """Parse commandline arguments."""
+
+    def __init__(self) -> None:
+        super().__init__(
+            prog="docsig",
+            formatter_class=lambda prog: _HelpFormatter(
+                prog, max_help_position=45
+            ),
+            description="Check signature params for proper documentation",
+        )
+        self._add_arguments()
+        self.args = self.parse_args()
+
+    def _add_arguments(self) -> None:
+        self.add_argument("-v", "--version", action="version", version="0.1.0")
+        self.add_argument(
+            "path",
+            nargs="+",
+            action="store",
+            type=_Path,
+            help="directories or files to check",
+        )
+        self.add_argument(
+            "-c",
+            "--check-class",
+            action="store_true",
+            help="check class docstrings",
+        )
+        self.add_argument(
+            "-D",
+            "--check-dunders",
+            action="store_true",
+            help="check dunder methods",
+        )
+        self.add_argument(
+            "-o",
+            "--check-overridden",
+            action="store_true",
+            help="check overridden methods",
+        )
+        self.add_argument(
+            "-p",
+            "--check-protected",
+            action="store_true",
+            help="check protected functions and classes",
+        )
+        self.add_argument(
+            "-P",
+            "--check-property-returns",
+            action="store_true",
+            help="check property return values",
+        )
+        self.add_argument(
+            "-i",
+            "--ignore-no-params",
+            action="store_true",
+            help="ignore docstrings where parameters are not documented",
+        )
+        self.add_argument(
+            "-a",
+            "--ignore-args",
+            action="store_true",
+            help="ignore args prefixed with an asterisk",
+        )
+        self.add_argument(
+            "-k",
+            "--ignore-kwargs",
+            action="store_true",
+            help="ignore kwargs prefixed with two asterisks",
+        )
+        self.add_argument(
+            "-n", "--no-ansi", action="store_true", help="disable ansi output"
+        )
+        self.add_argument(
+            "-S",
+            "--summary",
+            action="store_true",
+            help="print a summarised report",
+        )
+        self.add_argument(
+            "-s",
+            "--string",
+            action="store",
+            metavar="STR",
+            help="string to parse instead of files",
+        )
+
+
+if __name__ == "__main__":
+    _Parser()
+'''
+README_NO_COMMANDLINE_HELP = """
+arcon
+=====
+.. image:: https://img.shields.io/badge/License-MIT-yellow.svg
+    :target: https://opensource.org/licenses/MIT
+    :alt: License
+.. image:: https://img.shields.io/pypi/v/arcon
+    :target: https://pypi.org/project/arcon/
+    :alt: PyPI
+.. image:: https://github.com/jshwi/arcon/actions/workflows/build.yaml/badge.svg
+    :target: https://github.com/jshwi/arcon/actions/workflows/build.yaml
+    :alt: Build
+.. image:: https://github.com/jshwi/arcon/actions/workflows/codeql-analysis.yml/badge.svg
+    :target: https://github.com/jshwi/arcon/actions/workflows/codeql-analysis.yml
+    :alt: CodeQL
+.. image:: https://results.pre-commit.ci/badge/github/jshwi/arcon/master.svg
+   :target: https://results.pre-commit.ci/latest/github/jshwi/arcon/master
+   :alt: pre-commit.ci status
+.. image:: https://codecov.io/gh/jshwi/arcon/branch/master/graph/badge.svg
+    :target: https://codecov.io/gh/jshwi/arcon
+    :alt: codecov.io
+.. image:: https://readthedocs.org/projects/arcon/badge/?version=latest
+    :target: https://arcon.readthedocs.io/en/latest/?badge=latest
+    :alt: readthedocs.org
+.. image:: https://img.shields.io/badge/python-3.8-blue.svg
+    :target: https://www.python.org/downloads/release/python-380
+    :alt: python3.8
+.. image:: https://img.shields.io/badge/code%20style-black-000000.svg
+    :target: https://github.com/psf/black
+    :alt: Black
+.. image:: https://img.shields.io/badge/%20imports-isort-%231674b1?style=flat&labelColor=ef8336
+    :target: https://pycqa.github.io/isort/
+    :alt: isort
+.. image:: https://img.shields.io/badge/%20formatter-docformatter-fedcba.svg
+    :target: https://github.com/PyCQA/docformatter
+    :alt: docformatter
+.. image:: https://img.shields.io/badge/linting-pylint-yellowgreen
+    :target: https://github.com/PyCQA/pylint
+    :alt: pylint
+.. image:: https://img.shields.io/badge/security-bandit-yellow.svg
+    :target: https://github.com/PyCQA/bandit
+    :alt: Security Status
+.. image:: https://snyk.io/test/github/jshwi/arcon/badge.svg
+    :target: https://snyk.io/test/github/jshwi/arcon/badge.svg
+    :alt: Known Vulnerabilities
+.. image:: https://snyk.io/advisor/python/arcon/badge.svg
+  :target: https://snyk.io/advisor/python/arcon
+  :alt: arcon
+
+Persistent runtime config
+-------------------------
+
+Child class of ``argparse.ArgumentParser``
+
+Includes version argument as a default
+
+Default values are defined through pyproject.toml
+
+Includes additional argument adding methods
+
+.. code-block:: python
+
+    >>> __version__ = "0.1.0"
+    >>> from arcon import ArgumentParser
+
+Parsing comma separated list
+
+.. code-block:: python
+
+    >>> parser = ArgumentParser(__version__)
+    >>> parser.add_list_argument("-l", "--list")
+    >>> parser.parse_args(["--list", "comma,separated,list"])
+    Namespace(list=['comma', 'separated', 'list'])
+
+Parsing dict of comma separated lists
+
+.. code-block:: python
+
+    >>> parser = ArgumentParser(__version__)
+    >>> parser.add_dict_argument("-d", "--dict")
+    >>> parser.parse_args(["--dict", "key=comma,separated,list"])
+    Namespace(dict={'key': ['comma', 'separated', 'list']})
+"""