+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/pyaud-plugins/compare/v0.15.3...HEAD)
 ------------------------------------------------------------------------
+### Added
+- add `pyaud readme-help` plugin
 
 [0.15.3](https://github.com/jshwi/pyaud-plugins/releases/tag/v0.15.3) - 2023-08-04
 ------------------------------------------------------------------------