+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-[Unreleased](https://github.com/jshwi/pyaud-plugins/compare/v0.16.0...HEAD)
+[Unreleased](https://github.com/jshwi/pyaud-plugins/compare/v0.16.1...HEAD)
+------------------------------------------------------------------------
+
+[0.16.1](https://github.com/jshwi/pyaud-plugins/releases/tag/v0.16.1) - 2023-08-12
 ------------------------------------------------------------------------
 ### Fixed
 - ensure `pyaud readme-help` results are the same between versions